+++ data/sqlutil/resample_test.go
@@ -1,4 +1,4 @@
-package timeseries
+package sqlutil
 
 import (
 	"testing"
